requirements-docker.txt
asgiref==3.6.0
attrs==22.2.0
black==23.3.0
cffi==1.15.1
click==8.1.3
colorama==0.4.6
coverage==7.2.2
cryptography==39.0.1
Django==4.1.6
django-environ==0.9.0
django-filter==23.1
django-rest-knox==4.2.0
djangorestframework==3.14.0
exceptiongroup==1.1.0
iniconfig==2.0.0
MarkupSafe==2.1.2
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.1
Pillow==9.4.0
platformdirs==3.2.0
pluggy==1.0.0
psycopg2-binary==2.9.5
pycparser==2.21
pytest==7.2.1
pytest-django==4.5.2
python-decouple==3.8
pytz==2022.7.1
ruff==0.0.261
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.7
